Crate zng_wgt_text_input

Text input and label widgets.

Crate

This crate is part of the zng project.

Modules

label

Label text.

selectable

Selectable text.

Structs

DefaultStyle

W Text input default style.

FieldStyle

W Text input style that shows data notes, info, warn and error.

TextInput

W Simple text editor widget.

Statics

BASE_COLORS_VAR

Idle background dark and light color.

FIELD_HELP_VAR

Text shown under a FieldStyle when it has no data notes (no info, warn or error).

STYLE_FN_VAR

Contextual style variable.

Functions

base_colors

P Sets the colors used to compute all background and border colors in the text input style.

border_color

Default border color.

border_color_focused

Border color focused.

border_color_hovered

Border color hovered.

data_notes_adorner_fn

P Adorner property used by FieldStyle to show data info, warn and error.

default_context_menu

Context menu set by the DefaultStyle!.

default_popup_context_capture

Context captured for the context menu, set by the DefaultStyle!.

default_selection_toolbar

Selection toolbar set by the DefaultStyle!.

field_help

P Text shown under a FieldStyle when it has no data notes (no info, warn or error).

max_chars_count_adorner_fn

P Adorner property used by FieldStyle to show the count/max indicator.

style_fn

P Extends or replaces the widget style.